refactor: Split Edit into edit.go / edit_test.go
Move Edit, arrayHolder, and related helpers out of query.go.
Move Test_Edit, FuzzEdit, BenchmarkEdit out of query_test.go.
